requirements.txt
alembic
anthropic==0.76.0
anyio==4.8.0
async-timeout==4.0.3
biopython==1.84
boto3==1.34.2
botocore==1.34.2
celery==5.4.0
chardet==5.2.0
chromadb==0.4.22
crossrefapi==1.6.0
cryptography==44.0.2
defusedxml==0.7.1
docx2txt==0.9
ExtractTable==2.4.0
fastapi==0.115.8
flower==2.0.1
gevent
google-genai==1.60.0
google-search-results==2.4.2
gunicorn==22.0.0
httpx==0.28.1
json_repair==0.30.0
jupyter_kernel_client==0.7.1
jupyter_nbmodel_client==0.13.0
langchain==1.0.5
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-qdrant==1.1.0
langchain_anthropic==1.3.1
langchain_community==0.4.1
langchain_mongodb==0.11.0
langchain_openai==1.1.7
langfuse==3.9.1
langgraph==1.0.2
langgraph-checkpoint-postgres==3.0.1
langgraph-checkpoint-redis==0.3.4
llama-index==0.12.15
llmsherpa==0.1.4
lxml_html_clean==0.2.0
metapub==0.5.12
msoffcrypto-tool==5.4.2
nbconvert==7.16.6
numpy==1.26.2
onnxruntime==1.18.0
opencv-python>=4.9.0
openpyxl==3.1.5
pdf2doi==1.5.1
pika==1.3.2
pillow==10.2.0
psutil
psycopg-binary
psycopg-pool
psycopg2-binary
pydantic>=2.5.2,<3.0.0
pydantic-settings
PyJWT==2.10.0
pymongo==4.12.0
PyMuPDF==1.26.4
PyMuPDFb==1.23.7
pypdf==5.2.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-libsbml==5.20.2
python-multipart==0.0.6
PyYAML==6.0.1
rapidocr-onnxruntime
redis
regex
requests==2.32.5
scikit-learn==1.4.2
sentry-sdk[fastapi]
setproctitle
starlette==0.45.3
tenacity==8.5.0
tiktoken==0.7.0
tree-sitter==0.24.0
tree-sitter-language-pack
uvicorn==0.30.6
